WL#12108: Inject type cast nodes into the item tree to avoid data type mismatches
Makes the implicit CASTs involving DATE-related data types in basic
comparisons, explicit. This is achieved by allowing each operator to
adjust the arguments to the expected data-types. The adjustment is
limited to comparisons between fields and functions(so it avoids making
changes in comparisons with constants). The transformation is applied
towards the end of the optimization phase.

Reviewed-by: Dag Wanvik <dag.wanvik@oracle.com>
Change-Id: I7f7fc3cfcacccebf81b79e1823275bbd840f87dc