Bug #29684799 NDB_RESTORE_GCP.TEST FAIL TO CLEANUP
Reviewed-by: Mauritz Sundell <mauritz.sundell@oracle.com>

The ndb_restore_gcp mtr test was originally created to run
testBackup tests in mtr. These tests could not be run in
Autotest since they require the ndb_apply_status table, which is
not always present in Autotest runs. However, running the tests in
mtr causes backup files to be left behind, since the convention in
Autotest is to clean up backup files before each test, not after.

The ndb_restore_gcp testBackup tests are now moved to Autotest.
In testBackup, the epoch-restore functionality is modified to
handle a missing ndb_apply_status table. The epoch-restorer creates
ndb_apply_status through ndbapi if it does not exist.

The cleanup of backup files is also expanded to delete the local
copy of every backup. This local copy is created on every restore
by scp'ing or copying the backup files to the current working dir
for use by ndb_restore.

Change-Id: I8b58eb411f77118a19328431e19cafd3be94706c