BUG#29910397
When performing LCPs we call getCreateSchemaVersion and it goes to 0
at times for a short time. This makes LCP handling think the table
is being dropped and thus using the drop table logic. It can in effect
be in a TABLE_READ_ONLY state while ndb_restore is building indexes
offline. Fixed getCreateSchemaVersion to not change to schemaVersion
when state is TABLE_READ_ONLY.

Also fixed check in start_lcp_on_table.

Added numerous debug statements to track the changes of schema version
and table status in DBLQH.

Reviewed-by: Frazer Clement <frazer.clement@oracle.com>