WL #13000: Iterator UNION [patch 6/16; NULL rows from aggregation]
When setting up Item_fields with implicit aggregation (aggregation with no
GROUP BY), make sure to set the table row as nullable, so that Copy_field
is able to read NULL values out of even non-nullable fields if needed.
(This was in practice not needed before, as we'd usually copy Items and not
Fields, but it will be needed when we add ZeroRowsAggregatedIterator.)
This also fixes some bugs around CREATE TABLE, as demonstrated by the
test suite.

This exposed a related issue, in that Copy_field would assume that it could
alias an uchar* and bool*; ie., it assumed it could treat the table row's
NULL flag as a very simple NULL field bit mask. After we moved to my_bool to
bool, this is no longer allowed (although all the compilers we support happen
to have sizeof(bool) == 1, and thus it works by accident), and fixing it also
takes us one step closer to making TABLE::null_row private.

Change-Id: I116341b4f0f111ea028e2145961651afda1b8471