Bug #29620900 GCS MAY STILL DELIVER MESSAGES TO CLIENTS AFTER LEAVING THE GROUP
Description
========================================================================
GCS may still deliver messages to clients after leaving the group.
GCS does not check whether it still belongs to a group before delivering
a message to clients. The GCS engine thread that delivers messages is
not stopped until the client finalises the GCS interface, so GCS may
deliver messages between leaving the group and the client finalising the
GCS interface.

Solution
========================================================================
Check whether we still belong to a group before delivering a message.

RB: 22318
Reviewed-by: Tiago Jorge <tiago.jorge@oracle.com>
Reviewed-by: Venkatesh Venugopal <venkatesh.venugopal@oracle.com>