Bug 29823032 INPLACE UPGRADE 5.7.26 TO 8.0.16 FAILS WITH PARTITIONED TABLES AND LOWER_CASE_TABLE_NAMES=1
This bug is a regression after the fix for <Bug26925260 ERROR 1030 (HY000): GOT ERROR 44 - 'INNODB ERROR' FROM STORAGE ENGINE>
When set as lower_case_table_names=1 ,we should ideally convert all the identifiers in lower case
and store then in the disk in lower case, but this property is violated when it comes to naming
partitioned tables. Even with lower_case_table_names=1 when creating a partitioned table in case sensitive system
we store the partition suffixes (#P# and #SP#) in capitals and the partition names as specified in create table.
For example in case sensitive systems like UNIX if we create a partitioned table like

create table CUSTOMER ( id int ) partition by range (id) (partition CUST_ID_LESS_10 values less than(10)));
The file created in the disk is names as customer#P#CUST_ID_LESS_10.ibd wheres it should have been stored as
customer#p#cust_id_less.ibd.

Bug26925260 actually tried to fix this problem by converting the name to lower case when lctn=1 ,but this caused
regression when porting from 5.7 (where the name was in upper case) to 8.0 ( where we are trying to search
from lower case). We were unable to locate the files because of case difference and the upgrade failed. This has caused
many other regressions namely


Bug #29426720	WRONG MESSAGES IN ERROR-LOG WHEN USING PARTITION WITH LOWER-CASE-TABLE-NAMES=1
Bug #29627690	NOT ABLE TO IMPORT PARTITIONED TABLE SPACE OLDER THAN 8.0.14

FIX
---

We are reverting the bug fix for Bug 26925260. With this revert we are expecting some problems

1) With this bug fix upgrade will be successful from 5.7.* releases and releases prior to 8,0.14. (<=8.0.13)
2) Partitions created in  8.0.14-8.0.16 releases when upgraded to the version with this revert may face
   problems in alter rebuild. A quick workaround is to rename the table
3) Importing the tables from the affected versions [8.0.14-8.0.16]will not be possible.
4) Bug 26925260 will not be solved. This bug was for porting from windows to Linux.

Approved by: Mayank Prasad <mayank.prasad@oracle.com>