Bug#29897573: A FULL TABLE AGGREGATE QUERY IN DERIVED TABLE DOES NOT OFFLOAD Bug#29994087: SIG11 AT MATERIALIZEITERATOR::INIT() IN SQL/COMPOSITE_ITERATORS.CC
Some queries that contained subqueries were not offloaded to a
secondary storage engine. The optimizer tried to evaluate the
subqueries during optimization, and this failed because the secondary
engine could not return results until execution. This is fixed by
preventing evaluation of subqueries during optimization when using a
secondary storage engine.

Add a new query option that secondary storage engines can set to
prevent subqueries from being evaluated during optimization. The
option is checked before evaluating subqueries in the optimizer.

Setting of query options for secondary storage engines is removed from
open_secondary_engine_tables(). Instead, the engines can set the
options they need in their implementation of
handlerton::prepare_secondary_engine().

Item_bool_func2::update_used_tables() didn't properly accumulate
properties from the ESCAPE clause of a LIKE expression, so that it was
not detected that a LIKE expression contained a subquery in some
cases. A specialization of the function is added in Item_func_like to
make sure properties of the ESCAPE clause are accumulated in the LIKE
expression.

The test secondary_engine.no_constant_tables is removed, as it no
longer exercises the code that it was supposed to test. The code is
covered by tests in other test suites.

Change-Id: I945df19452cbca0d5f5621f0bb68351009c9e79c