Bug#29925009: ADD AN OPTION TO GENERATE GDB INDEX
Add a CMake option to enable generation of a .gdb_index section in the
binaries, which makes it faster to load them in a debugger. The option
is called ADD_GDB_INDEX. It defaults to ON in debug builds on Linux
when the LLD linker is used, and to OFF otherwise. The option has no
effect if another linker than LLD or GNU gold is used.

Change-Id: I845446f6e8b566884cbdb4329ea1b593fbf17671