Bug#29931153 slow tests on windows
Change
======

Limit the wait_interval time for ProcessLauncher::read() to 100ms max

Problem
=======

ProcessLauncher::read() reads from a pipe in non-blocking way and
waits between tries to create read-with-timeout.

The wait_interval is calculated as:

   wait_interval = min(timeout / 10, 1ms)

which leads to 3sec wait_interval for 30sec timeout.

That means, if pipe isn't closed on first check, the function will wait
3sec before trying again.

RB: 22380