Bug#29939331 - THE NEGATION OF A "<=>" COMPARISON MALFUNCTIONS DEPENDING ON THE COLUMN'S TYPE
Corrected wrong NULL handling during folding.

The <=> operator is "NULL safe", that is, it treats NULL as a value
rather than "unknown" in the SQL three value logic sense.
Therefore, if we find that a comparison using it cannot possibly be
true due to the constant's value, we should fold it to FALSE, even if
the field is nullable.

Change-Id: Ibf862a5819e9509c241974682b748b0e8d6f71af