Bug #29934590	DISTRO BUILD FAILS IN INNODB CODE
When using `strncpy`, one has to keep in mind, that if the source string is as long
or longer than the destination buffer, then there will be no `\0` appended.
Thus, one has to make sure to always append the `\0` themselves, if they want the
destination buffer to contain a `\0`-terminated string.

Reviewed-by: Jusufadis Bakamovic <jusufadis.bakamovic@oracle.com>
RB:22398