Bug#30019632: ERROR LOG REPORTING DELAYED DURING INNODB START-UP WHEN UNNECESSARY
During start-up, we buffer error log messages and flush them later.

In Phase I, "before command-line options are handled", because we
do not yet know how to format timestamps, what file to log to (if
any), and so on.

In Phase II, "after options handling, but before loadable error
log services are available" (or, for all intents and purposes,
during InnoDB start-up / DD-upgrade), because we may not be able
to log the way the DBA requested if that would rely on a loadable
log service we can't yet load.

If Phase II takes very long, we'll eventually reach a point where
not giving any feedback to the user is worse than giving it in an
undesired format (i.e. using the built-in default logger rather
than a loadable logger that may be configured). Before this patch,
we essentially gave Inno an arbitrary amount of time hoping it
would complete start-up, and if it didn't, we started flushing
error log entries in the default format so the DBA would know what's
going on. The reason we did not flush immediately was that the logger
could be configured to write to stderr (rather than a file we know
the name of), and in that case writing first one format and then the
other might break logging tools.

This patch improves that mechanism so that we ONLY delay the flushing
during Inno start-up if a) we write to a stream, not a named file,
AND b) a non-default log-error-services pipeline is used.
This means we only delay flushing in the one scenario where flushing
immediately can be unsafe; if we don't have multiple log-writers,
it's always safe, and if we do have multiple log-writers that write
to separate files, it's also safe. In these latter two cases, which
should constitute the vast majority of configurations, the events
will be logged immediately during this stage, which should constitute
a quality-of-life improvement both for DBAs and InnoDB-developers.

Approved by: Jens Even <jens.even.blomsoy@oracle.com>