Bug#29949096 ERROR FOR FAILURE TO MIGRATE DURING DISCOVER
No error is set when trying to discover an unsupported(ie. too old)
table from NDB. The ndbcluster plugin will just push some warnings and
return an error from ndbcluster_discover(). However the calling
convention for this function is that an error should be set before
returning.

This cause a mysqld crash after failure to migrate during discover with
fix for BUG#29934763. Earlier the failed discover was retried and the
retry caused a different code path to be used which set the error,
although lots of duplicated warnings was produced which indicated that
something was wrong already at that time.

Fix by returning ER_NO_SUCH_TABLE and remove all the duplicated
warnings.

Change-Id: I66942d027b4d3ffc66ca98c4f8119c68e41fb71f