Bug #29534647 ANALYZE TABLE DOES NOT CONCLUDE, AND STUCK ON "SYSTEM LOCK"
MySQL Cluster index statistics are calculated based on the
topology of one fragment of an ordered index.

The fragment chosen in any particular index is decided at index
creation time - both when the index is created, and when a node or
system restart recreated the index locally.

The calculation is partly based on the number of fragments in the
index, and this can change when a table is reorganized, so that on
the next node restart, the restarting node makes a different choice
of chosen fragment, resulting in none, one or two fragments being
used to generate index statistics, resulting in errors for the
NdbApi operation, and the ANALYZE TABLE command which uses it.

This is solved by modifying the online table reorganization to
recalculate the chosen fragment immediately, so that all nodes
are aligned before and after any subsequent restart.

With this change :
  Initially :
    An index statistics fragment A is chosen
    Index statistics generated from this fragment are generated
     and served

  At the end of a table reorganisation :
    The fragment chosen for generating index statistics is
     recalculated by all nodes based on the new number of
     index fragments (Fragment B)
    The index statistics for fragment A continue to be used.

  At any further node restarts
    The fragment chosen for generating index statistics is
     recalculated by the node, based on the new number of
     index fragments (Fragment B)

  On a subsequent ANALYZE TABLE operation
    New index statistics are generated based on the topology
     of Fragment B
    The old index statistics based on the topology of Fragment
     A are deleted

A new testcase : ndb_addnode_analyze has some coverage of this
area.

Approved by : Martin Skold <martin.skold@oracle.com>