Bug#29871932 WAIT FOR SYNC IN SECOND SERVER
Problem:
--------

The ndb_dd_backuprestore test involves restoring a backup and
then waiting for the auto sync functionality to synchronize the
DD. The metadata_change_detection_interval was reduced in one of
the MySQL servers to hasten auto sync but this was omitted in
the other. As a result, it was likely that some metadata objects
existed in the DD of 1 MySQL server and not the other. This led
to the second MySQL server issuing warnings during later DDL
statements

Fix:
----

- Reduce the interval and wait for sync in the second MySQL server
as well before proceeding further
- The different wait loops i.e. wait for changes to be detected,
wait for changes to be synced, and wait for sync blacklist to be
empty have been extracted into their own include files

Change-Id: I311bcbbe38bb57a3ccf6ccf56bfb42c7b3148a32