BUG#29900621: GR_PRIMARY_MODE_GROUP_OPERATIONS_41_1 FAILS SPORADICALLY
Test gr_primary_mode_group_operations_41_1 is failing
sporadically during a primary switch due to a incorrect number of
pending transactions to finish before the primary switch, in this
case a 2 is expected but a 3 is returned.
In some corner cases, the statement that is used to check this value
can be accounted as a pending transaction, which explains the +1
value.

The test case now takes that into consideration.

ReviewBoard: 22441