WL#12453 New configuration in NDB
The old configuration format in NDB is limited to handling at most 16381 sections.
Since for example each pair of nodes that connects to each other will need a
communication section the maximum number of data nodes supported would be less
than 128.

The new configuration layout will be able to handle up to 4G sections.
This will be able to support several thousands of nodes.

Only send communication sections to nodes that they are involved in. This
will ensure that the size of the configuration sent to nodes is kept within
reasonable limits. It now grows linearly with the number of nodes and not
quadratically as before. Quadratic growth will give lots of headaches when
reaching many hundreds and even more at thousands of nodes.

Fix test results for ndb_config and ndb_config2. Order of
nodes in printouts differ now, now in nodeid order, before
first nodes for MGM servers, next for MySQL Servers and
next data nodes and nodeid order within each group.

Reviewed-by: Mauritz Sundell <mauritz.sundell@oracle.com>
Reviewed-by: Mikael Ronström <mikael.ronstrom@oracle.com>
Reviewed-by: Sanjana DS <sanjana.ds@oracle.com>