Bug#29909905 MAIN.OPTIMIZER_DEBUG_SYNC FAILS ON PB2
The last testcase in this test simulates an out-of-memory
condition. However, since the server is multithreaded, any thread
might terminate due to this condition.

Since this isn't possible to handle in a meningful way, we just delete
that part of the test.

Also, this push re-enables the test in Valgrind.

Change-Id: Ib1157883cf27cedd89e96f7261aef1325373edc0