Bug#30042677 unknown option -fno-inline with /Ob0 on clang-cl
Problem
=======

clang-cl supports building with sanitizers like -DWITH_UBSAN on windows.

To make debugging problems easier the mysql-cmake disables inlining
via '-fno-online' which is supported by both clang and gcc.

On windows the clang-cl driver isn't aware of that option and complains
for every file that's compile that the unknown option is ignored.

Change
======

- use /Ob0 instead of -fno-inline when using clang-cl.

RB: 22572