Bug#29965125 WL #12680: MGMCONFIG-T UNIT TEST FAILING IN PB2-ASAN
Problem
=======
Default configuration sections did not have the configuration section type
set when unpacked into memory.

That caused a memory leak since the destructor of section did not destroy
the entries of the section in that case.

Solution
========
Set correct section type also when unpacking default sections.

Also added a suppression for LSAN such that it did not check for memory
leaks in /bin/bash, which also caused test failures.

Reviewed-by: Sanjana DS <sanjana.ds@oracle.com>