Bug#29990693 VALGRIND ERROR: CONSTANT FOLDING W/EXECUTE MIGHT LEAD TO BUG
The error happens when constant folding decides to do constant folding
of a function's arguments, cf. method fold_arguments. The third
argument to fold_condition is "Item **retcond", i.e. a pointer to a
condition that might be simplifed as a result of calling
fold_condition.

This happens in our repro. Now, when this happens as a normal
execution, the change of condition is done with a plain assignment,
and the logic in fold_arguments does the right thing: it detects this
change and updates the argument in question with the new Item. Thus,
we have two values being changed:

1) *retcond  provided by fold_arguments (by fold_condition)
2) func->arguments()[i] (by fold_arguments).

But during PS + EXECUTE, the changes are not done with plain
assignments, they are performed using THD::change_item_tree, which
registers the change for rollback after execution.

Now, change 2) above is good. But change 1) registers an item tree
change of a stack variable, which is no longer live when the rollback
happens, i.e. the rollback writes the original pointer value somewhere
on the runtime stack, presumably unused memory since the rollback is a
cleaning up action with a low number of stack frames.

Fix: Remove the helper stack variable. This makes rolling back safe.

No extra test is added, the existing const_folding.test showed the
issue before the fix when run with valgrind.

Note: this bug was discovered after commit of Bug#29939331, which
added a test which provoked the error, but is not the cause of it.

Change-Id: I1b5c83f6e60eb194217f2e164ee5f75798f094d4