BUG#29996547 NDB_STORED_USER is not propagated over MySQL replication
A new test case ndb_rpl.ndb_rpl_stored_grants describes and tests
the interaction between NDB distributed grants and MySQL replication.

Any ACL change executed on a binlog-enabled server will be written
to the binlog. This means that changes to users with NDB_STORED_USER
will be replicated on the slave cluster, and also that changes to
purely local users on the binlogging server will be copied to the
replication slave endpoint.

The actual fix for the bug is a small change in
Ndb_local_connection::run_acl_statement() so that instead of calling
execute_query_iso(), which prevents statements from being written
to the binlog, it directly calls execute_query().

The new test case is disabled pending a fix for bug#29966540.

Change-Id: I366b579fd6e5520dba962d35f761cf4a6f499955