BUG#29892598: GROUP_REPLICATION.GR_CONCURRENT_UNINSTALL_START FAILS ON PB2
Class Message_service_handler contains a thread to handle incoming
messages, that thread termination was signaling too soon that it
was terminated.
That was leaving dangling pointers which could be hit when the
commands
  UNINSTALL PLUGIN group_replication;
  INSTALL PLUGIN group_replication SONAME group_replication.so;
where executed within a short time window.

To fix the above situation, Message_service_handler::dispatcher()
now signals that the thread terminated after its memory is released.
The Message_service_handler::terminate() always waits for the
signal that thread terminated.

ReviewBoard: 22502