Bug#29910978 SCHEMA SYNC TAKES TOO LONG WITH VALGRIND
Problem:
--------

Schema synchronization took a sizable amount of time especially
on slower hosts. On investigation, it was discovered that fetching
the list of tables belonging to an engine was the bottleneck. The
fetch_schema_components() function instantiates each dd::Table
which caused the issue

Fix:
----

Use the new DD API functions added as part of the fix for
Bug#29906844 DD FUNCTION FOR FETCHING TABLES IN ENGINE. These
functions return the table names directly. Additionally, the
NdbApi equivalent function of obtaining table names belonging
to a schema has been tweaked to skip the NDB utility tables
which are marked as hidden in the DD. The synchronization of
these utility tables are handled as special cases by the binlog
thread and thus are of no interest during auto sync and "normal"
schema synchronization

Change-Id: I99acbea2c89ce0e84b278a48b1a70101885106dc