Bug#27160203 - SELECT A NODE FOR TC PLACEMENT SHOULD IGNORE NODES NOT STARTED
Problem: Sometimes, an unavailable node (that hasn't started yet, or
the system is in single user mode) could be selected as a TC
node. This is fixed by WL#10172. However, selecting a node in a
round-robin fashion is not addressed.

Fix: Check whether the node is available for the round-robin selection
case as well.

Approved-by: Mikael Ronström <mikael.ronstrom@oracle.com>