WL#12966 Replication with privilege checks
Description
-----------
Make it possible to enable privilege checks on replication channels, so that it
is possible to replicate securely from untrusted masters.

It's possible to restrict the replication applier by specifying the privilege
checking context in which it can execute. Specifically, it's possible to
configure a user account through `CHANGE MASTER TO PRIVILEGE_CHECKS_USER =
[user | NULL]`.

Contributors
------------
Designed-by: Sven Sandberg <sven.sandberg@oracle.com>

Developed-by: Abhinav Agarwal <abhinav.ab.agarwal@oracle.com>
Developed-by: Daogang Qu <bill.qu@oracle.com>
Developed-by: Neha Kumari <neha.n.kumari@oracle.com>
Developed-by: Pedro Figueriedo <pedro.figueiredo@oracle.com>

Reviewed-by: Harin Vadodaria <harin.vadodaria@oracle.com>
Reviewed-by: Luís Soares <luis.soares@oracle.com>
Reviewed-by: Pedro Gomes <pedro.gomes@oracle.com>
Reviewed-by: Pedro Ribeiro <pedro.a.ribeiro@oracle.com>
Reviewed-by: Prasad Vasudevan <prasad.vasudevan@oracle.com>