Bug #29882690	UNDETECTED DEADLOCK WITH GAP AND INSERT INTENTION LOCKS IS POSSIBLE
This patch moves the task of detecting a dealock cycle among data locks from the transaction thread, to a dedicated background thread.
This patch also fixes Bug #21278148 DEADLOCKS MISSED BY INFORMATION_SCHEMA.INNODB_METRICS LOCK_DEADLOCKS COUNTER

The idea is that for each waiting transaction, we store a pointer to a transaction which is blocking it.
(Note, that in general there might be multiple such transactions, but we pick any of them).
We let the dedicated deadlock detector thread to peek the contents of these pointers for transansactions which are sleeping in slots.
This creates a small snapshot of a subset of the big conceptual wait-for-graph, restricted only to blocked transactions, and only to
one edge per transaction.
This does not even have to be a consistent snapshot, as the only thing we want to find in it are deadlock cycles, which are stable
by definition.
As the graph has out-degree=1, it has a shape of forests possibly leading to cycles, and thus finding those cycles is trivial.
All this can be done almost lock-free, as long as care is taken not to derefernce pointers without making sure that trx is still in the slot.
For this purpose we store reservation_no on each slot, and compare current slot->reservation_no to the one in the snapshot.

Reviewed-by: Debarun Banerjee <debarun.banerjee@oracle.com>
RB:22374