BUG#30020127 TEST BINLOG_REPLICATION_APPLIER_PRIV_CHECKS FAILS
Description
-----------
Test `binlog.binlog_replication_applier_priv` is failing when executed with
`GTID_MODE=ON`.

Analysis
--------
When `GTID_MODE=ON`, the execution of `BINLOG` statement holding a
`Write_rows_event` fails, when executing it after a `Format_description_event`
has just been applied. This happens because the `Format_description_event`
places `GTID_NEXT` in an undertermined state to be properly set by applying
either an `Anonymous_gtid_log_event` or a `Previous_gtid_log_event` - which is
what is expected to be applied after a `Format_description_log_event`.

In the mentioned test script case, a simpler case is being tested, that is, the
case where `FILE` and `INSERT` privileges are checked against
`Format_description_log_event` and `Write_rows_log_event` respectively.

Fix
---
Move the test to the `binlog_nogtid` suite since it doesn't impact the case
being tested and prevents the test from being executed with `GTID_MODE=ON`.

Reviewed-by: Pedro Gomes <pedro.gomes@oracle.com>